moved test_base.py to ./tests and into a unittest
