+++ amaranth/sim/_vcdwriter.py
@@ -24,7 +24,7 @@ class _VCDWriter(Observer):
         return format.format(value).expandtabs().replace(" ", "_")
 
     def __init__(self, state, design, *, vcd_file, gtkw_file=None, traces=(), **kwargs):
-        super.__init__(**kwargs)
+        super().__init__(**kwargs)
         self.state = state
 
         # Although pyvcd is a mandatory dependency, be resilient and import it as needed, so that
